% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Choose whether to stay or switch the contestant door choice and return the final door choice.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value indicating whether the contestant stays with their initial door choice (\code{TRUE} by default) or switches (\code{FALSE}).}

\item{opened.door}{An integer representing the door that was opened to reveal a goat (1, 2, or 3).}

\item{a.pick}{An integer representing the contestant's initial door choice (1, 2, or 3).}
}
\value{
The function returns a number between 1 and 3, indicating the final door choice. With the stay strategy, the final choice will be the same as the initial choice. With the switch strategy, the final choice will be the door that was not the initial pick and not the door opened by the host.
}
\description{
\code{change_door()} returns the final door choice based on whether the contestant switches or stays with the original door.
}
\details{
This is the third step of the game where the contestant chooses to switch or stay with their door.
}
\examples{
  change_door( stay=T, opened.door, a.pick )
  change_door( stay=F, opened.door, a.pick )
}
