% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether the contestant has won based on the final door choice.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer representing the contestant's final door choice (1, 2, or 3).}

\item{game}{A character vector of length 3 representing the Monty Hall game setup. One door has "car" and the other two have "goat".}
}
\value{
The function returns "WIN" if the final door has the car or "LOSE" if the final door has a goat.
}
\description{
\code{determine_winner()} returns "WIN" if the final door has the car or "LOSE" if the final door has a goat.
}
\details{
This labels the game as a win or loss.
}
\examples{
  determine_winner( final.pick, game )
}
