% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Select a goat door from the two remaining doors.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector of length 3 representing the Monty Hall game setup. One door has "car" and two doors have "goat".}

\item{a.pick}{An integer representing the contestant's initial door choice (1, 2, or 3).}
}
\value{
The function returns a number between 1 and 3, indicating the host door choice. The door number will be a door that is not the contestant choice and not the car.
}
\description{
\code{open_goat_door()} selects one of the two remaining doors that does not have the car.
}
\details{
This is the second step of the game where the host selects a door with a goat behind it.
}
\examples{
  open_goat_door( this.game, my.initial.pick )
}
