% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Combines all of the functions to play a full game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a dataframe that shows if the contestant will "WIN" or "LOSE" based on the two strategy choices.
}
\description{
\code{play_game()} runs a simulation of the game and creates a dataframe to show if the contestant wins or loses based on the strategy chosen.
}
\details{
This brings all of the steps together into a simulation of the game.
}
\examples{
  play_game( )
}
