% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple games.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{An integer specifying the number of games to play. The default is 100.}
}
\value{
The function returns a dataframe that shows how many times the contestant will "WIN" or "LOSE" based on the two strategy choices.
}
\description{
\code{play_n_games()} runs multiple simulations of the game and create a dataframe to show how many times the contestant wins or loses based on the strategy chosen.
}
\details{
This runs multiple simulations of the game and returns the results in a single dataframe.
}
\examples{
  play_n_games( n=100 )
}
