% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Randomly select a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3, indicating the contestant door choice.
}
\description{
\code{select_door()} randomly selects one of the three doors.
}
\details{
This is the first step of the game where the contestant selects a door.
}
\examples{
  select_door()
}
